% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{createChebiLink}
\alias{createChebiLink}
\title{ChEBI link}
\usage{
createChebiLink(cas, name)
}
\arguments{
\item{cas}{CAS reference ID for drug}

\item{name}{Molecule name}
}
\value{
Link string which can be put into a datatable
}
\description{
Create link to CAS name search in ChEBI
}
\examples{
createChebiLink("drugID", "drugname")

}

