% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{fixWells}
\alias{fixWells}
\title{Fix well names}
\usage{
fixWells(wellString)
}
\arguments{
\item{wellString}{}
}
\description{
Pad column numbers in well names so A1 becomes A01
but A16 stays A16
}

