% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{\%w/o\%}
\alias{\%w/o\%}
\title{Without}
\usage{
x \%w/o\% y
}
\arguments{
\item{x}{vector of values}

\item{y}{vector of values to be removed from \code{x}}
}
\value{
vector of remaining values
}
\description{
Handy function to remove existing items from a vector - borrowed from the
documentation for  \code{\link{match}}
}
\examples{
1:8 \%w/o\% c(1,4)

# character vectors
letters[1:8] \%w/o\% letters[c(1,4)]
}

