% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import_functions.R
\name{importDispensingFileEcho550}
\alias{importDispensingFileEcho550}
\title{Import Echo 550 dispensing file}
\usage{
importDispensingFileEcho550(inputPath)
}
\arguments{
\item{inputPath}{Path to a dispensing file}
}
\value{
List containing dispensing data in position 1,
extra information in position 2, with an md5 checksum appenedd as the final column
}
\description{
Import the
}
\examples{
tmp <- importDispensingFileEcho550(inputPath)
}

