% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import_functions.R
\name{importPlateReaderDataEnvision}
\alias{importPlateReaderDataEnvision}
\title{Import plate reader data}
\usage{
importPlateReaderDataEnvision(inputPath, plateRowCount = 16,
  plateColCount = 24)
}
\arguments{
\item{inputPath}{Path to file}

\item{plateRowCount}{number of rows on plate, defined alphabetically (caps) from A: default=16}

\item{plateColCount}{number of columns on plate, defined numerically from 01: default=24}
}
\value{
list with data and meta data
}
\description{
Import data for plate 1 from an Envision plate reader
}
\examples{
tmp <- importPlateReaderDataEnvision(inputPath)
}

