% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import_functions.R
\name{listImportedChecksums}
\alias{listImportedChecksums}
\title{List imported file checksums}
\usage{
listImportedChecksums(dbPath)
}
\arguments{
\item{dbPath}{path to database}
}
\value{
data frame of imported checksums in column 1 and imported=TRUE in column 2
}
\description{
Get the checksums for all the
}

