% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import_functions.R
\name{matchFilenameToTable}
\alias{matchFilenameToTable}
\title{Match filename to database table}
\usage{
matchFilenameToTable(fileList, fileClassifiers)
}
\arguments{
\item{fileList}{vector of input files}

\item{fileClassifiers}{data frame of file classifiers and related database tables}
}
\description{
Match filename to database table
}
\examples{
matchFilenameToTable(fileList, fileClassifiers)
}

