% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import_functions.R
\name{putChosenFilesIntoDatabase}
\alias{putChosenFilesIntoDatabase}
\title{Put data from chosen file into correct database table}
\usage{
putChosenFilesIntoDatabase(fileList, chosenIndividual, fileClassifiers, dbPath,
  pathToZipFile, tmpdir = tempDirectory)
}
\arguments{
\item{chosenIndividual}{input$dropdownSampleGroups}
}
\description{
Put data from chosen file into correct database table
}
\examples{
putChosenFilesIntoDatabase(fileList, chosenIndividual, fileClassifiers, dbPath, pathToZipFile, tmpdir=tempDirectory)

}

