% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions_sqlite.R
\name{queryDB}
\alias{queryDB}
\title{Query database}
\usage{
queryDB(dbPath, SQLstring)
}
\arguments{
\item{dbPath}{Path to the database}

\item{SQLstring}{SQL string defining query to be performed}
}
\value{
A data frame (tbl_df) containing the query results
}
\description{
Run an SQL query string on the database
}
\examples{
results <- queryDB(dbPath, SQLstring)
}

