% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions_sqlite.R
\name{readTableFromDB}
\alias{readTableFromDB}
\title{Read table from database}
\usage{
readTableFromDB(dbPath, tableName)
}
\arguments{
\item{dbPath}{Path to the database}

\item{tableName}{Name of database table to read from}
}
\value{
A data frame (tbl_df) containing the table
}
\description{
Read full contents of a single table from the database
}
\examples{
results <- readTableFromDB(dbType, dbPath, tableName)
}

