% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions_sqlite.R
\name{writeTableToDB}
\alias{writeTableToDB}
\title{Write table to database}
\usage{
writeTableToDB(dbPath, tableName, inboundData)
}
\arguments{
\item{dbPath}{Path to the database}

\item{tableName}{Name of database table to append to}

\item{inboundData}{Data frame to be appended to table}
}
\value{
single Boolean value for matching of column names and types
between the input data frame and the target database table
}
\description{
write a data frame to the database, appending it to an existing sqlite table
}
\examples{
writeTableToDB(dbType, dbPath, tableName, inboundData)
}

