% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling_functions.R
\name{zipFileContents}
\alias{zipFileContents}
\title{Get Zip File Contents}
\usage{
zipFileContents(pathToZipFile, tmpdir = tempDirectory)
}
\arguments{
\item{pathToZipFile}{}
}
\value{
dataframe with all csv files and their checksums
}
\description{
Get a list of the files included in the zip
}
\examples{
zipFileContents(pathToZipFile)

}

